+++ examples/gnu-cat.rs
@@ -6,6 +6,7 @@ use flood_tide::NameVal;
 use flood_tide::Opt;
 use flood_tide::OptNum;
 use flood_tide::OptParseError;
+use flood_tide::HelpVersion;
 
 //----------------------------------------------------------------------
 //{{{ TEXT
@@ -99,7 +100,18 @@ struct CmdOptConf {
     pub flag_number: bool,
     pub flag_squeeze: bool,
     //
+    pub flag_help: bool,
+    pub flag_version: bool,
+    //
     pub arg_params: Vec<String>,
+}
+impl HelpVersion for CmdOptConf {
+    fn is_help(&self) -> bool {
+        self.flag_help
+    }
+    fn is_version(&self) -> bool {
+        self.flag_version
+    }
 }
 
 //----------------------------------------------------------------------
@@ -258,6 +270,8 @@ mod example {
             " flag_number_nb: false,",
             " flag_number: false,",
             " flag_squeeze: false,",
+            " flag_help: false,",
+            " flag_version: false,",
             " arg_params: [\"f1\", \"-\", \"f2\"] }"
         );
         assert_eq!(thing, expect);
@@ -287,6 +301,8 @@ mod example {
             " flag_number_nb: false,",
             " flag_number: false,",
             " flag_squeeze: false,",
+            " flag_help: false,",
+            " flag_version: false,",
             " arg_params: [\"f1\", \"-\", \"f2\"] }"
         );
         assert_eq!(thing, expect);
@@ -335,6 +351,8 @@ mod example {
             " flag_number_nb: false,",
             " flag_number: false,",
             " flag_squeeze: false,",
+            " flag_help: false,",
+            " flag_version: false,",
             " arg_params: [\"f1\", \"-\", \"f2\"] }"
         );
         assert_eq!(thing, expect);